refactor(framework): keep the profile subpath lightweight
Move the pure subset math (subsetStandardManifest, ownedExtensionsForExcludedModules,
SubsetOptions) into manifest.ts so profile.ts no longer imports create-app.ts.
create-app.ts statically imports frameworkComposition (composition-lazy.ts ->
zod-openapi + module runtime), so the lightweight
@voyant-travel/framework/profile subpath — used by the source-free managed admin
for snapshot math — was pulling the full runtime graph (a pre-existing regression,
since profile.ts already imported subsetStandardManifest from create-app.ts).

- manifest.ts now owns the subset primitives (pure: manifest data + the
  lightweight findCapabilityGaps graph engine only). Co-located with the
  capability graph they validate against.
- create-app.ts, profile.ts, index.ts, and the tests import them from manifest.js.
  profile.ts's dependency graph no longer reaches create-app.ts / composition-lazy.ts.
- Renamed create-app.test.ts -> manifest.test.ts to match the code under test.
