+++ packages/framework/src/manifest.ts
@@ -18,7 +18,7 @@
  * Workstream B of the consolidated-deployments RFC: the standard registry's
  * factories live in this package alongside the manifest (the "which + order").
  */
-import type { CapabilityGraph } from "@voyant-travel/hono/composition"
+import { type CapabilityGraph, findCapabilityGaps } from "@voyant-travel/hono/composition"
 
 export interface FrameworkManifest {
   modules: readonly string[]
@@ -163,3 +163,94 @@ export const FRAMEWORK_EXTENSION_OWNERSHIP = {
   (typeof FRAMEWORK_RUNTIME_MANIFEST.extensions)[number],
   readonly (typeof FRAMEWORK_RUNTIME_MANIFEST.modules)[number][]
 >
+
+/** Options for {@link subsetStandardManifest}. */
+export interface SubsetOptions {
+  /** Specifiers to remove entirely (rejected if unknown, `isRequired`, or depended-on). */
+  exclude?: readonly string[]
+}
+
+/**
+ * The standard extensions owned by any of the excluded specifiers — an extension
+ * whose declared owner module (see `FRAMEWORK_EXTENSION_OWNERSHIP`) is being
+ * removed. Excluding a module must cascade to these, or the removed surface
+ * partially leaks: e.g. dropping `bookings` while `finance/bookings-create-extension`
+ * (mounting under `/v1/admin/bookings`) stays mounted (voyant#2104). Ownership is
+ * declared, not name-matched, so path-mounted extensions cascade correctly.
+ */
+export function ownedExtensionsForExcludedModules(excluded: Iterable<string>): string[] {
+  const excludedSet = excluded instanceof Set ? excluded : new Set(excluded)
+  const owned: string[] = []
+  for (const extension of FRAMEWORK_RUNTIME_MANIFEST.extensions) {
+    const owners = FRAMEWORK_EXTENSION_OWNERSHIP[extension]
+    if (owners.some((owner) => excludedSet.has(owner))) owned.push(extension)
+  }
+  return owned
+}
+
+/**
+ * Apply `exclude` to the standard set (ADR-0007), returning the module/extension
+ * specifiers that should mount. Pure and provider-free (manifest math only, no
+ * runtime composition), so it is unit-testable, reusable by tooling (`db doctor`),
+ * and safe to import from the lightweight `@voyant-travel/framework/profile`
+ * subpath. Throws — fail-loud at boot, never a runtime 500 — when `exclude` names
+ * a specifier absent from the standard set (a typo), names an `isRequired` module,
+ * or leaves a still-mounted module's `requires` unmet (drop the consumers too).
+ */
+export function subsetStandardManifest({ exclude = [] }: SubsetOptions = {}): {
+  modules: string[]
+  extensions: string[]
+} {
+  const excludeSet = new Set(exclude)
+
+  if (excludeSet.size > 0) {
+    const known = new Set<string>([
+      ...FRAMEWORK_RUNTIME_MANIFEST.modules,
+      ...FRAMEWORK_RUNTIME_MANIFEST.extensions,
+    ])
+    const unknown = [...excludeSet].filter((spec) => !known.has(spec)).sort()
+    if (unknown.length > 0) {
+      throw new Error(
+        `createVoyantApp: exclude names ${unknown.length} specifier(s) not in the standard set: ` +
+          `${unknown.join(", ")}. Only standard framework modules/extensions can be excluded.`,
+      )
+    }
+
+    const graph: CapabilityGraph = FRAMEWORK_CAPABILITY_GRAPH
+    const required = [...excludeSet].filter((spec) => graph[spec]?.isRequired).sort()
+    if (required.length > 0) {
+      throw new Error(
+        `createVoyantApp: cannot exclude required module(s): ${required.join(", ")}. ` +
+          "They are foundational and cannot be removed.",
+      )
+    }
+  }
+
+  // Cascade: dropping a module also drops the standard extensions it owns. They
+  // mount under the module's surface, so leaving them would partially leak the
+  // removed surface. Ownership is declared (`FRAMEWORK_EXTENSION_OWNERSHIP`), so
+  // path-mounted extensions cascade by declaration rather than an unsound
+  // name-match (voyant#2104). Idempotent — already-excluded extensions no-op.
+  for (const extension of ownedExtensionsForExcludedModules(excludeSet)) {
+    excludeSet.add(extension)
+  }
+
+  const modules = FRAMEWORK_RUNTIME_MANIFEST.modules.filter((m) => !excludeSet.has(m))
+  const extensions = FRAMEWORK_RUNTIME_MANIFEST.extensions.filter((e) => !excludeSet.has(e))
+
+  // The capability graph is validated over what actually mounts: dropping a
+  // module a still-mounted module depends on — without also dropping the consumer
+  // — fails loudly here rather than as a runtime 500.
+  const gaps = findCapabilityGaps(modules, FRAMEWORK_CAPABILITY_GRAPH)
+  if (gaps.length > 0) {
+    const detail = gaps
+      .map((g) => `"${g.capability}" (required by ${g.requiredBy.join(", ")})`)
+      .join("; ")
+    throw new Error(
+      `createVoyantApp: exclude leaves unmet capabilities: ${detail}. ` +
+        "Exclude the consumers too (capability replacement is a future release).",
+    )
+  }
+
+  return { modules, extensions }
+}
